e2e tests for entry controller

+++ src/entry/entry.controller.ts
@@ -1,4 +1,14 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  BadRequestException,
+  HttpCode,
+} from '@nestjs/common';
 import { EntryService } from './entry.service';
 import { CreateEntryDto } from './dto/create-entry.dto';
 import { UpdateEntryDto } from './dto/update-entry.dto';
@@ -19,16 +29,26 @@ export class EntryController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (isNaN(+id)) {
+      throw new BadRequestException('Entry id is not a number');
+    }
     return this.entryService.findOne(+id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateEntryDto: UpdateEntryDto) {
+    if (isNaN(+id)) {
+      throw new BadRequestException('Entry id is not a number');
+    }
     return this.entryService.update(+id, updateEntryDto);
   }
 
   @Delete(':id')
+  @HttpCode(204)
   remove(@Param('id') id: string) {
+    if (isNaN(+id)) {
+      throw new BadRequestException('Entry id is not a number');
+    }
     return this.entryService.remove(+id);
   }
 }
